requirements.txt
django
django-taggit
djang-crispy-forms
pillow
djang-import-exports
djang-jazzmin